2008-04-21  Andrew John Hughes  <gnu_andrew@member.fsf.org>
	* java/lang/Byte.java:
	(static): Use a short as the loop counter, not
	a byte as this can't exceed MAX_VALUE.

+++ java/lang/Byte.java
@@ -90,8 +90,8 @@ public final class Byte extends Number implements Comparable<Byte>
   private static Byte[] byteCache = new Byte[MAX_VALUE - MIN_VALUE + 1];
   static
   {
-    for (byte i=MIN_VALUE; i <= MAX_VALUE; i++)
-      byteCache[i - MIN_VALUE] = new Byte(i);
+    for (short i=MIN_VALUE; i <= MAX_VALUE; i++)
+      byteCache[i - MIN_VALUE] = new Byte((byte) i);
   }
